Added : Abscence modal + put all modal in a folder
