+++ front-end/src/App.js
@@ -11,7 +11,7 @@ import { Dashboard } from "./Pages/Staff/Dashboard";
 import { Calendar } from "rsuite";
 import { SelectInput } from "./components/FormInput";
 import Modal from "./components/Modal";
-import SuggestActivityForm from "./components/NewActivity";
+import SuggestActivityForm from "./components/modals/NewActivity";
 
 function App() {
 
@@ -40,7 +40,7 @@ function App() {
 
             {/* Modal is here because it is called by the sidebar */}
             <Modal isOpen={isModalOpen_SuggestActivity} onClose={closeModal_SuggestActivity}>
-               <SuggestActivityForm/>
+               <SuggestActivityForm Name={"IBRAAAAA"} Reason={"Lebron James"}/>
             </Modal>
 
             {/*  If there is need for another modal in the slidebar, it will be used like this
