+++ front-end/src/components/modals/LaunchActivity.js
