+++ front-end/src/style/modals/Absence.css
@@ -1,0 +1,37 @@
+.AbsenceBody{
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    width: 30vw;
+    height: 70vh;
+}
+
+.AbsenceName{
+    position: absolute;
+    top: 50px;
+    left: 50px;
+}
+
+.H2Absence{
+    margin-bottom: 30px;
+}
+
+.AbsenceReason{
+    width: 25vw;
+    height: 20vh;
+    overflow: scroll;
+    padding: 10px;
+
+    border-style: solid;
+    border-width: 2px ; 
+    border-color: #999999;
+    border-radius: 5px;
+}
+
+.ButtonAbsence{
+    margin: 30px 10px 20px 10px;
+    padding: 3% 5%;
+    border-radius: 5px;
+    font-weight: 500;
+}
