+++ front-end/src/style/modals/LaunchActivity.css
