+++ front-end/src/style/modals/NewActivity.css
@@ -9,28 +9,34 @@
     height: 70vh;
 }
 
-.H2NewActivity{
+.NewActivityTitle{
     margin-bottom: 30px;
 }
 
-.ActivityForm input,.ActivityForm select,.ActivityForm TagPicker{
+.ActivityForm input, .ActivityForm select{
     margin: 0 10px 20px 10px;
     padding: 3% 5%;
     border-radius: 5px;
-    border: 1px solid #ccc;
+    border: 2px solid #ccc;
     font-weight: 500;
+    background-color: #FFFFFF;
+    color: #000000;
 }
 
-.ActivityType{
+.ActivityForm .ActivityType{
     margin: 0 10px 20px 10px;
     padding: 2% 4%;
     border-radius: 5px;
-    border: 1px solid #ccc;
+    border: 2px solid #ccc;
 }
-.TagPicker{
+.ActivityForm .TagPicker{
     margin: 0 10px 20px 10px;
     border-radius: 5px;
     border: 1px solid #ccc;
+}
+
+.SubmitButton{
+    padding: 2% 5% !important;
 }
 
 .Column{
