+++ front-end/src/style/modals/Refusal.css
